backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
cryptography==43.0.0
pikepdf==9.0.0
pydantic==2.9.0
httpx==0.27.0
python-whois==0.9.4
beautifulsoup4==4.12.3
